+++ src/Portal.tsx
@@ -1,13 +1,12 @@
 import { CProps } from '../types';
 import useIsBrowser from './hooks/useIsBrowser';
 import { element } from './utils';
-import { ReactNode } from 'react';
+import { ElementType, ReactNode } from 'react';
 import { createPortal } from 'react-dom';
 
-type Props = CProps & { root?: string; children: ReactNode };
+type Props = { root?: string; children: ReactNode };
 
-export default function Portal({ root = 'body', ...rest }: Props) {
+export default function Portal<TTag extends ElementType>({ root = 'body', ...rest }: Props & CProps<TTag>) {
   const isBrowser = useIsBrowser();
-
   return isBrowser ? createPortal(element(rest), document.querySelector(root) as HTMLElement) : null;
 }