+++ tests/Portal.test.tsx
@@ -1,15 +1,15 @@
 import { Portal } from '../src';
 import { render } from '@testing-library/react';
 
-test('Portal', () => {
+test('Check Portal', () => {
   const { getByText, getByTestId } = render(
     <div data-testid="container">
-      <Portal data-testid="element">I'm being portalled</Portal>
+      <Portal>I'm being portalled</Portal>
     </div>,
   );
 
   // Check the element render
-  const element = getByTestId('element');
+  const element = getByText("I'm being portalled");
   expect(element).toBeDefined();
 
   // Make sure the element don't render inside it's parent component